Refactor project structure. Add javadocs for classes public classes. Javadocs for public methods will be added in a follow-up commits.
