+++ src/main/java/core/data/Data.java
@@ -1,0 +1,8 @@
+package core.data;
+
+/**
+ * Data is storing some piece of information that is used and transmitted throughout framework.
+ */
+public class Data {
+    public Object data;
+}
