+++ src/main/java/core/pipeline/PipelineBlock.java
@@ -1,11 +1,19 @@
-package model.pipeline;
+package core.pipeline;
 
-import model.data.Data;
+import core.data.Data;
 
 import java.util.ArrayList;
 import java.util.List;
 
-public  abstract class PipelineBlock<TInput extends Data, TOutput extends Data> {
+/**
+ * Block is used as a core part of pipeline.
+ *
+ * <p> Pipeline consists of different blocks that are sending data to each other.
+ *
+ * @param <TInput>  input that is expected in a block to be received from other PipelineBlocks in the pipeline
+ * @param <TOutput> output that is produced by PipelineBlock and propagated to listening blocks in pipeline.
+ */
+public abstract class PipelineBlock<TInput extends Data, TOutput extends Data> {
 
 
     protected List<PipelineBlock<TOutput, ?>> listeners;
@@ -20,7 +28,7 @@ public  abstract class PipelineBlock<TInput extends Data, TOutput extends Data> 
         this.listeners.add(listener);
     }
 
-    public void notifyListeners(PipelineBlock<?, TOutput> sender, TOutput data){
+    public void notifyListeners(PipelineBlock<?, TOutput> sender, TOutput data) {
         listeners.forEach(pipelineBlock -> pipelineBlock.InputDataReady(sender, data));
     }
