+++ src/main/java/core/pipeline/impl/AlgorithmHostBlock.java
@@ -1,13 +1,18 @@
-package core.pipeline;
+package core.pipeline.impl;
 
 
 import core.algorithm.Algorithm;
-import model.data.Data;
-import model.pipeline.PipelineBlock;
+import core.pipeline.PipelineBlock;
+import core.data.Data;
 
 /**
- * FYI
- * block not necessary has to have algorithm
+ * A {@link PipelineBlock}, that is storing and executing {@link Algorithm} and
+ * transferring its results to other blocks.
+ *
+ * <p>Block does not necessary has to have algorithm.
+ *
+ * @param <TInput>  is an input type of the {@link Algorithm}, hosted by this Block.
+ * @param <TOutput> is an output type of the {@link Algorithm}, hosted by this Block.
  */
 public class AlgorithmHostBlock<TInput extends Data, TOutput extends Data> extends PipelineBlock<TInput, TOutput> {
 
@@ -22,7 +27,7 @@ public class AlgorithmHostBlock<TInput extends Data, TOutput extends Data> exten
 
     @Override
     public void InputDataReady(PipelineBlock<?, TInput> sender, TInput tInput) {
-        if(enabled) {
+        if (enabled) {
             TOutput result = algorithm.execute(tInput);
             notifyListeners(this, result);
         }
