+++ src/main/java/core/pipeline/impl/SequentialPipeline.java
@@ -1,10 +1,11 @@
-package core.pipeline;
+package core.pipeline.impl;
 
-import model.data.Data;
-import model.pipeline.PipelineBlock;
+import core.pipeline.PipelineBlock;
+import core.data.Data;
 
 import java.util.List;
 
+// TODO(avlomakin): add javadoc here
 public class SequentialPipeline {
 
     public SequentialPipeline(List<PipelineBlock> blocks) {
