+++ src/main/java/features/numInfoFromImage/ImageSeriesVoxelVolumeCalcAlg.java
@@ -1,12 +1,15 @@
-package core.algorithm.impl.num;
+package features.numInfoFromImage;
 
 import core.algorithm.Algorithm;
-import model.data.Data;
-import model.data.ImageSeries;
-import model.data.MedImage;
+import core.data.Data;
+import core.data.medImage.ImageSeries;
+import core.data.medImage.MedImage;
 
 import java.util.UUID;
 
+/**
+ * Calculates the volume of the given {@link ImageSeries}.
+ */
 public class ImageSeriesVoxelVolumeCalcAlg implements Algorithm<ImageSeries, Data> {
 
     private final UUID ONE_PIXEL_VOLUME_ID = UUID.fromString("34091644-e39a-11e8-9f32-f2801f1b9fd1");
@@ -19,7 +22,7 @@ public class ImageSeriesVoxelVolumeCalcAlg implements Algorithm<ImageSeries, Dat
 
         double result = 0;
         for (MedImage image: input.images) {
-            //how can i do it properly?
+            // TODO(sabrinamusatian): rewrite it using Arrays.stream
             for (byte[] line: image.getImagePixels()) {
                 for (byte aLine : line) result += (aLine != 0) ? onePixelVolume : 0;
             }
