--- src/main/java/model/data/Data.java
@@ -1,9 +1,0 @@
-package model.data;
-
-/**
- * Created by sabrina on 28.10.2018.
- */
-public class Data {
-    public Integer hash;
-    public Object data;
-}
